% Generated by roxygen2 (4.0.2): do not edit by hand
\name{add_scholar_cites_to_bib}
\alias{add_scholar_cites_to_bib}
\title{Fetch citation counts for references with google scholar ids}
\usage{
add_scholar_cites_to_bib(author_id, bibin, bibout = NULL, clean = TRUE)
}
\description{
Fetch citation counts for references with google scholar ids
}
\examples{
\dontrun{
add_scholar_cites_to_bib("cuXoCA8AAAAJ", 'mypubs.bib')
}
}

