% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ailinkedfiles}
\alias{ailinkedfiles}
\title{pure R implementation of ailinkedfiles.pl}
\usage{
ailinkedfiles(x)
}
\arguments{
\item{x}{Path to Illustrator file (ai,pdf)}
}
\value{
Character vector of paths to linked files
}
\description{
This is much slower than perl version for large files
This appears to be due to speed of readLines
}
\author{
jefferis
}
\seealso{
Other linked_from: \code{\link{linked_from_ai}}
}

