% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bib.r
\name{biber_cache}
\alias{biber_cache}
\alias{remove_biber_cache}
\title{Return full path to the biber (biblatex) cache or delete it to fix problems}
\usage{
biber_cache()

remove_biber_cache()
}
\value{
\code{biber_cache} returns a character vector

\code{remove_biber_cache} returns a logical indicating success.
}
\description{
\code{biber_cache} returns the path to the cache

\code{remove_biber_cache} deletes cache to fix bib problems.
}
\details{
Biblatex is a flexible alternative to bibtex for processing bibliographies.
The most commonly used tool for processing biblatex is biber. Unfortunately
biber has a bug that frequently results in failure to produce a bibliography.
see
\url{http://tex.stackexchange.com/questions/140814/biblatex-biber-fails-with-a-strange-error-about-missing-recode-data-xml-file}
for discussion.
}
\seealso{
Other bib: \code{\link{bibdesk_clean}};
  \code{\link{lyx_citekeys}}
}

