% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paperutils-package.r
\name{dr_paperutils}
\alias{dr_paperutils}
\title{Check current set up of external tools used by paperutils}
\usage{
dr_paperutils()
}
\description{
Check current set up of external tools used by paperutils
}
\examples{
dr_paperutils()
}
