% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fix_sla_links}
\alias{fix_sla_links}
\title{Re-link all linked files in Scribus file in relative manner}
\usage{
fix_sla_links(x)
}
\arguments{
\item{x}{path to Scribus file.}
}
\description{
Re-link all linked files in Scribus file in relative manner
}

