% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdfbookmarks.r
\name{extract_bookmarks}
\alias{extract_bookmarks}
\alias{insert_bookmarks}
\title{Insert and Extract bookmarks from a PDF}
\usage{
extract_bookmarks(pdfin, bookmarks)

insert_bookmarks(pdfin, bookmarks, pdfout = pdfin)
}
\arguments{
\item{pdfin, pdfout}{Input and Output PDFs}

\item{bookmarks}{Path to a bookmarks PDF file in pdftk format or a character
vector of length >1 containing bookmark text.}
}
\value{
location of pdftk bookmark file

location of modified PDF
}
\description{
Insert and Extract bookmarks from a PDF
}
\seealso{
\code{\link{make_bookmark_text}}
}
\author{
jefferis
}
