% Generated by roxygen2 (4.0.2): do not edit by hand
\name{is.pptx}
\alias{is.pptx}
\title{Test if file is in pptx format}
\usage{
is.pptx(x, Verbose = TRUE)
}
\arguments{
\item{x}{Path to PowwerPoint pptx file}

\item{Verbose}{Whether to provide a user message on error.}
}
\description{
Test if file is in pptx format
}

