% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lyxutils.r
\name{lyx_citekeys}
\alias{lyx_citekeys}
\title{List the bibtex citekeys for a LyX document}
\usage{
lyx_citekeys(auxfile = NULL)
}
\arguments{
\item{auxfile}{Path to a latex aux file (see details for default behaviour)}
}
\value{
A character vector of citation keys in the order in which they are
  cited in the LyX document.
}
\description{
Uses latex \code{.aux} file to extract citation keys
}
\details{
The required input is not the LyX file itself, but rather the 
  auxfile produced by a latex/bibtex run (note this is also produced by 
  biblatex). When \code{auxfile=NULL} the \code{\link{current_lyx_tempfile}} 
  is used to determine the appropriate auxfile.
  
  auxfiles do seem to have a variety of different formats for recording the
  citekey list - this function has been tested for files processed using 
  \code{natbib} + \code{bibdesk}.
}
\seealso{
Other LyX: \code{\link{current_lyx_tempfile}},
  \code{\link{lyxexport}}, \code{\link{lyxhtml2html}}

Other bib: \code{\link{bibdesk_clean}},
  \code{\link{biber_cache}}
}
\concept{LyX}
\concept{bib}
