% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lyxutils.r
\name{lyxexport}
\alias{lyxexport}
\title{Export a LyX file to another format (by default xhtml)}
\usage{
lyxexport(x, outfile = NULL, format = "xhtml")
}
\arguments{
\item{x}{The input file}

\item{outfile}{Path to output file (defaults to <infilestem>.<format>)}

\item{format}{The LyX shortname for the format (see details)}
}
\description{
Export a LyX file to another format (by default xhtml)
}
\details{
Look in Tools->Preferences->File Handling->File Formats->Short Name 
to see which parameter (which differs from the format name in the
File->Export menu) should be passed as format.
}
\seealso{
Other lyx: \code{\link{current_lyx_tempfile}},
  \code{\link{lyx_citekeys}}, \code{\link{lyxhtml2html}}
}

