% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lyxutils.r
\name{lyxhtml2html}
\alias{lyxhtml2html}
\title{Convert a LyXHTML file to an html file that can be opened by Word}
\usage{
lyxhtml2html(infile, outfile = NULL)
}
\arguments{
\item{infile}{path to xhtml file}

\item{outfile}{path to output html file (defaults to <infilestem>.<html>)}
}
\description{
LyXHTML is currently (LyX 2.1 series) the highest quality html export
available by default. However, it cannot be opened by word immediately
because it interposes a small xml type header before the html content. This
function removes that header and renames the file with a .html extension by
default.
}
\examples{
\dontrun{
lyxfile=system.file('tests/testthat/testdata/lyx/test.lyx', package='paperutils')
lyxhtml=tempfile(pattern = basename(lyxfile), fileext = '.xhtml')
lyxexport(lyxfile, outfile=lyxhtml)
lyxhtml2html(lyxhtml)
}
}
\seealso{
Other LyX: \code{\link{current_lyx_tempfile}},
  \code{\link{lyx_citekeys}}, \code{\link{lyxexport}}
}
