% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{paperutils}
\alias{paperutils}
\alias{paperutils-package}
\title{paperutils: PDF manipulation utilities (wrapping pdftk and ghostscript)}
\description{
paperutils: PDF manipulation utilities (wrapping pdftk and ghostscript)
}
\section{External dependencies}{


See the \href{https://github.com/jefferis/paperutils}{README}.
}

\section{Package options}{


  There is one package option at present: \itemize{

  \item paperutils.convert can be used to set the path to the ImageMagick
  convert function used by \code{\link{convert_pptx_pdfs}} and friends.

  }
}
\seealso{
\code{\link{convert_pptx_pdfs}}
}

