% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pdf2png}
\alias{pdf2png}
\title{Convert pdf file to a png using imagemagick convert tool}
\usage{
pdf2png(pdf, png, res = 300)
}
\arguments{
\item{pdf}{Path to input pdf}

\item{png}{Path to output png}

\item{res}{Output resolution in dpi}
}
\value{
Logical value indicating success (or failure) of operation
}
\description{
Convert pdf file to a png using imagemagick convert tool
}

