% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdftk.r
\name{pdftk}
\alias{pdftk}
\title{Return absolute path to pdftk binary}
\usage{
pdftk(mustWork = TRUE)
}
\arguments{
\item{mustWork}{Whether to error out if tool cannot be found (default
\code{TRUE})}
}
\description{
Return absolute path to pdftk binary
}
