% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdftotext.R
\name{pdftotext}
\alias{pdftotext}
\title{Invoke pdftotext tool to extrat text from a PDF document}
\usage{
pdftotext(x, outfile = NULL, first = NULL, last = NULL, layout = FALSE,
  table = FALSE, raw = FALSE, ...)
}
\arguments{
\item{x}{Path to pdf}

\item{outfile}{Optional output file (defaults to <pdfstem.txt>)}

\item{first}{Optional page to start conversion}

\item{last}{Optional page to finish conversion}

\item{layout, table, raw}{pdftotext options (all default \code{FALSE})}

\item{...}{additional arguments passed to \code{readLines} when
\code{outfile=FALSE}}
}
\value{
the return value of the system call or, if outfile has the signalling
  value of \code{FALSE}, the contents of the text file as a character vector.
}
\description{
Invoke pdftotext tool to extrat text from a PDF document
}
\details{
Run pdftotext -help for details
}
