% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdftotext.R
\name{pdftotextbin}
\alias{pdftotextbin}
\title{Return absolute path to pdftotext binary}
\usage{
pdftotextbin(mustWork = TRUE)
}
\arguments{
\item{mustWork}{Whether to error out if tool cannot be found (default
\code{TRUE})}
}
\description{
Return absolute path to pdftotext binary
}
