% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib.r
\name{read_bibdesk_bib}
\alias{read_bibdesk_bib}
\title{Read a BibDesk bib file after cleaning it to remove problem fields}
\usage{
read_bibdesk_bib(x)
}
\arguments{
\item{x}{Path to a BibDesk bib file}
}
\value{
See \code{\link[RefManageR]{ReadBib}}
}
\description{
Read a BibDesk bib file after cleaning it to remove problem fields
}
\examples{
\dontrun{
r=read_bibdesk_bib("~/Greg/ProfessionalAdmin/cv_lyx/JefferisPublications.bib")
citekeys=unname(names(r))
citekeys
}
}
