% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{read_bibdesk_static_groups}
\alias{read_bibdesk_static_groups}
\title{Read in the citekeys for each static group in a BibDesk file}
\usage{
read_bibdesk_static_groups(x)
}
\arguments{
\item{x}{Path to a BibDesk bib file}
}
\value{
A named list of vectors (see details)
}
\description{
Read in the citekeys for each static group in a BibDesk file
}
\details{
The function returns a named list with one element for each static
  group. Each element contains the bibtex citekey for each item in the
  corresponding static group.
}
\examples{
\donttest{
res=read_bibdesk_static_groups("~/Greg/ProfessionalAdmin/cv_lyx/JefferisPublications.bib")
res$Primary
}
}
