% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sla_linked_files}
\alias{sla_linked_files}
\title{List all linked files from given Scribus file (sla)}
\usage{
sla_linked_files(x)
}
\arguments{
\item{x}{path to Scribus file.}
}
\value{
Character vector of paths to linked files.
}
\description{
List all linked files from given Scribus file (sla)
}

