% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptxutils.r
\name{unzip_pptx}
\alias{unzip_pptx}
\title{Unzip a PowerPoint pptx to a temporary directory}
\usage{
unzip_pptx(x, exdir = tempfile(pattern = basename(x)))
}
\arguments{
\item{x}{Path to PowwerPoint pptx file}

\item{exdir}{Path to temporary diectory}
}
\value{
\code{exdir}
}
\description{
Unzip a PowerPoint pptx to a temporary directory
}

