% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptxutils.r
\name{zip_pptx_dir}
\alias{zip_pptx_dir}
\title{Zip up a temporary directory into a PowerPoint pptx file}
\usage{
zip_pptx_dir(x, pptx, action = c("freshen", "update", "error"),
  files = NULL)
}
\arguments{
\item{x}{Path to directory containing expanded PowerPoint data}

\item{pptx}{Path to new output pptx}

\item{action}{What to do if \code{pptx} points to an existing file}

\item{files}{Optional character vector of files to freshen (when
\code{action='freshen'})}
}
\description{
Zip up a temporary directory into a PowerPoint pptx file
}
\details{
Note that the temporary directory itself does not form part of the 
  resultant zip folder, which only includes the 3 folders underneath it 
  (_rels, docProps, ppt)
}

