% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cv.R
\name{bibdesk_clean}
\alias{bibdesk_clean}
\title{Clean up a bibtex file produced by BibDesk}
\usage{
bibdesk_clean(bibin, bibout = NULL)
}
\arguments{
\item{bibin}{Path to input file}

\item{bibout}{Path to output file. The defacult value of \code{NULL} will
generate a temporary file.}
}
\value{
A character vector containing the path to the output file or
  NA_character_ when conversion fails
}
\description{
Uses the command line \bold{bibtool} program with a custom resource file to
remove over-long fields that can cause problems for bibtex parsers. These
include the \code{annote} and \code{bdsk-url-*} and \code{bdsk-file} fields.
}
\examples{
rsc_file=system.file("bibtool","bibdesk-clean.rsc", package = 'paperutils')
}
\seealso{
Other bib: \code{\link{biber_cache}},
  \code{\link{remove_biber_cache}};
  \code{\link{lyx_citekeys}}
}

