% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bibtool}
\alias{bibtool}
\title{Run bibtool command line utility}
\usage{
bibtool(infile, ..., cmds = NULL, outfile = NULL)
}
\arguments{
\item{infile,outfile}{Input and Output files}

\item{...}{Additional character arguments passed straight to bibtool
executable}
}
\description{
Run bibtool command line utility
}
\examples{
\dontrun{
# remove the Annote field
bibtool('in.bib', outfile='in-cleaned.bib', 'delete.field Annote')
}
}

