% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdftk.r
\name{extract_pdf}
\alias{extract_pdf}
\alias{extractpdf}
\alias{extractpdf}
\title{Extract pages from a PDF to new PDF(s)}
\usage{
extract_pdf(pdfin, pages, pdfout, prefix = NULL, DryRun = F,
  gscomp = FALSE, bookmarks = NULL)

extractpdf(pdfin, pages, pdfout, prefix = NULL, DryRun = F,
  gscomp = FALSE, bookmarks = NULL)
}
\arguments{
\item{pdfin, pdfout}{Path to input and output PDF}

\item{pages}{Integer vector or list. If named, the names specify output files}

\item{prefix}{a prefix to add to the start of all output PDF file names}

\item{DryRun}{Just say what would happen (when TRUE)}

\item{gscomp}{Compress with ghostscript}

\item{bookmarks}{Bookmarks info file (pdftk format) that can be used to add
bookmarks to the output PDF}
}
\description{
Extract pages from a PDF to new PDF(s)
}
\examples{
\dontrun{
aux='/path/to/myfile.aux'
pdffile='/path/to/myfile.pdf'
figurepages=findfigs(aux)
extract_pdf(pdffile,figurepages,prefix='LastAuthor_',gscomp=TRUE)
}
}
\seealso{
\code{\link{gscompress}}
}
\author{
jefferis
}
