% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slautils.r
\name{fix_sla_links}
\alias{fix_sla_links}
\title{Convert linked files in Scribus file from absolute to relative paths}
\usage{
fix_sla_links(x, writeout = FALSE)
}
\arguments{
\item{x}{path to Scribus file.}

\item{writeout}{logical indicating whether to write out fixed Scribus file or
character vector specifying path for file.}
}
\value{
character vector of fixed paths (invisibly when \code{writeout=TRUE})
}
\description{
Convert linked files in Scribus file from absolute to relative paths
}
\details{
when \code{writeout=TRUE} the existing file is fixed and
  overwritten.
}

