% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs.r
\name{gscompress}
\alias{gscompress}
\title{Compress a pdf using ghostscript}
\usage{
gscompress(pdfin, pdfout, firstpage, lastpage, pdflevel = "1.5",
  gsopts = "", Force = TRUE)
}
\arguments{
\item{pdfin, pdfout}{Input and Output pdfs}

\item{firstpage, lastpage}{Page range (defaults to firstpage = 1, lastpage = last in pdf)}

\item{pdflevel}{Minium compatible pdf version of output file}

\item{gsopts}{Further options for ghostscript}

\item{Force}{When FALSE, only update file for newer input}
}
\value{
character vector with path to pdfout
}
\description{
When pdfout is missing it defaults to <input>_gso.pdf
}
\details{
This seems to work well for pdflatex output
}
\examples{
\dontrun{
gscompress('statistics.pdf',gsopts='-dSubsetFonts=true -dEmbedAllFonts=true')
}
}
\author{
jefferis
}

