% Generated by roxygen2 (4.0.1): do not edit by hand
\name{linked_from_sla}
\alias{linked_from_sla}
\title{List all linked files from given Scribus file (sla)}
\usage{
linked_from_sla(x)
}
\arguments{
\item{x}{path to Scribus file.}
}
\value{
Character vector of paths to linked files.
}
\description{
List all linked files from given Scribus file (sla)
}

