% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pdftk.r
\name{numpages}
\alias{numpages}
\title{Return number of pages in pdf}
\usage{
numpages(pdf)
}
\arguments{
\item{pdf}{Path to input pdf file}
}
\value{
number of pages
}
\description{
Return number of pages in pdf
}
\author{
jefferis
}

