% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{paperutils}
\alias{paperutils}
\alias{paperutils-package}
\title{paperutils: PDF manipulation utilities (wrapping pdftk and ghostscript)}
\description{
paperutils: PDF manipulation utilities (wrapping pdftk and ghostscript)
}

