% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptxutils.r
\name{pngres}
\alias{pngres}
\title{Return (horizontal) resolution in dots per inch of a png file}
\usage{
pngres(x)
}
\arguments{
\item{x}{Path to a png file}
}
\value{
numeric vector giving resolution in dots per inch
}
\description{
Return (horizontal) resolution in dots per inch of a png file
}

